src/modules/db_operations/requirements.txt
Flask==1.1.2
Flask_SQLAlchemy==2.5.1
logzero==1.3.1
python-dotenv==0.19.1
itsdangerous==2.0.1
sentry-sdk==1.5.4
pg8000
PyYAML>=4.2b1